requirements.txt
requests
pandas==2.3.2
scikit-learn==1.6.1
google-cloud-vision==3.10.2
google-auth==2.40.3
pydub==0.25.1
st-annotated-text==4.0.2